up/back-port the travel/fuel balance changes

+++ src/main/java/dev/amble/ait/tardis/util/TardisUtil.java
@@ -3,7 +3,7 @@ package dev.amble.ait.tardis.util;
 import java.util.*;
 import java.util.function.Predicate;
 
-import dev.amble.ait.tardis.handler.FuelHandler;
+import dev.amble.ait.tardis.v2.handler.FuelHandler;
 import dev.amble.lib.data.CachedDirectedGlobalPos;
 import dev.amble.lib.data.DirectedBlockPos;
 import dev.amble.lib.util.TeleportUtil;
@@ -484,7 +484,7 @@ public class TardisUtil {
     }
 
     public static double estimatedFuelCost(PlayerEntity player, Tardis tardis, double distance){
-        double speed = Math.max(tardis.travel().speed(), 1);
+        int speed = Math.max(tardis.travel().speed(), 1);
         double ticksRequired = distance / speed;
         double perTick = FuelHandler.getPerTickFuelCost(speed, tardis.travel().instability());
         return perTick * ticksRequired;
